Revert "Move internal headers to a subdirectory (#507)"
This reverts 4700f87f6472deaa9239764acafbb8eed470ca18
