+++ docs/Doxyfile
@@ -784,9 +784,7 @@ WARN_AS_ERROR          = YES
 # spaces. See also FILE_PATTERNS and EXTENSION_MAPPING
 # Note: If this tag is empty the current directory is searched.
 
-INPUT                  = ../library/include/rocsolver \
-                         ../library/include/rocsolver/internal \
-                         ../library/src/include/ideal_sizes.hpp
+INPUT                  = ../library/include/rocsolver ../library/src/include/ideal_sizes.hpp
 
 # This tag can be used to specify the character encoding of the source files
 # that doxygen parses. Internally doxygen uses the UTF-8 encoding. Doxygen uses
