+++ library/CMakeLists.txt
@@ -6,6 +6,6 @@
 set(rocsolver_SOVERSION 0.1)
 
 # Create version header from templated .in file using CMake info
-configure_file(include/rocsolver/internal/rocsolver-version.h.in "${PROJECT_BINARY_DIR}/include/rocsolver/internal/rocsolver-version.h")
+configure_file(include/rocsolver/rocsolver-version.h.in "${PROJECT_BINARY_DIR}/include/rocsolver/rocsolver-version.h")
 
 add_subdirectory(src)
