+++ library/include/rocsolver/rocsolver-extra-types.h
