+++ library/include/rocsolver/rocsolver-version.h.in
