+++ library/include/rocsolver/rocsolver.h
@@ -9,12 +9,11 @@
 #ifndef ROCSOLVER_H
 #define ROCSOLVER_H
 
-/* clang-format off */
-#include "internal/rocsolver-export.h"
-#include "internal/rocsolver-extra-types.h"
-#include "internal/rocsolver-aliases.h"
-#include "internal/rocsolver-functions.h"
-#include "internal/rocsolver-version.h"
-/* clang-format on */
+/* library headers */
+#include "rocsolver-aliases.h"
+#include "rocsolver-export.h"
+#include "rocsolver-extra-types.h"
+#include "rocsolver-functions.h"
+#include "rocsolver-version.h"
 
 #endif /* ROCSOLVER_H */
